epub-maker: split cover auto renderers into svg and raster modules
